# swagger-sample.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Healthier Food Choices"
  },
  "servers": [
    {
      "url": "https://scenic-virgin-islands-80550.herokuapp.com/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "List all healthier food choices",
        "tags": [
          "Root"
        ],
        "responses": {
          "200": {
            "description": "An array of healthier food choices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": {
                      "main_food_group": "Beverages",
                      "company_name": "Aero Vending Private Limited",
                      "product_name": "Aero Vending Apple Concentrate",
                      "brand_name": "Aero Vending",
                      "product_weight": "5 US Gallon"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/main_food_group": {
      "get": {
        "summary": "List only the available food groups",
        "tags": [
          "main_food_group"
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "example": [{
                      "main_food_group": "Seafood",
                      "company_name": "Clouet Trading Pte Ltd",
                      "product_name": "Ayam Brand Chilli Tuna Light",
                      "brand_name": "Ayam Brand",
                      "product_weight": "185g"
                    },
                    {
                      "main_food_group": "Seafood",
                      "company_name": "Clouet Trading Pte Ltd",
                      "product_name": "Ayam Brand Club Sardines in Extra Virgin Olive Oil",
                      "brand_name": "Ayam Brand",
                      "product_weight": "120g"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new food group",
        "tags": [
          "main_food_group"
        ],
        "responses": {
          "201": {
            "description": "Null response"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/main_food_group/{group}": {
      "get": {
        "summary": "List all the available food of a specified group",
        "tags": [
          "main_food_group"
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "description": "The group of the food to retrieve, e.g. 'beverages', 'cereal', 'crisps', etc.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "main_food_group": "Seafood",
                      "company_name": "Clouet Trading Pte Ltd",
                      "product_name": "Ayam Brand Chilli Tuna Light",
                      "brand_name": "Ayam Brand",
                      "product_weight": "185g"
                      },
                      {
                      "main_food_group": "Seafood",
                      "company_name": "Clouet Trading Pte Ltd",
                      "product_name": "Ayam Brand Club Sardines in Extra Virgin Olive Oil",
                      "brand_name": "Ayam Brand",
                      "product_weight": "120g"
                      }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific item",
        "tags": [
          "main_food_group"
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "description": "The food group of the resource to update",
            "schema": {
              "type": "object"
            }
          }
        ],
        "requestBody": {
          "description": "Updated fields for resource",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                		"main_food_group": "Beverages",
                		"company_name": "New Company Name",
                		"product_name": "New Product Name",
                		"brand_name": "New brand name",
                		"product_weight": "999 litres"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                  		"main_food_group": "Beverages",
                  		"company_name": "New Company Name",
                  		"product_name": "New Product Name",
                  		"brand_name": "New brand name",
                  		"product_weight": "999 litres"
                    },
                    {
                      "main_food_group": "Beverages",
                      "company_name": "Aero Vending Private Limited",
                      "product_name": "Aero Vending Apple Concentrate",
                      "brand_name": "Aero Vending",
                      "product_weight": "5 US Gallon"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a main_food_group and the associated items",
        "tags": [
          "main_food_group"
        ],
        "parameters": [
          {
            "name": "group",
            "in": "path",
            "required": true,
            "description": "The food group of the resource to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "example": 
                    {
                  		"message":"{main_food_group} deleted"
                    }
                }
              }
            }
          }
        }
      }
    }
  }
}